[quant][graphmode] Rename backend_config_dict folder to backend
Summary:
att

Test Plan:
```
python test/fx2trt/test_quant_trt.py
```
Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
