+++ torch/ao/quantization/fx/__init__.py
@@ -1,4 +1,4 @@
 from .prepare import prepare
 from .convert import convert
 from .fuse import Fuser
-from .backend_config_dict import get_tensorrt_backend_config_dict
+from .backend import get_tensorrt_backend_config_dict
